+++ lib/src/runner.dart
@@ -3,7 +3,6 @@ import 'package:dartle/dartle.dart';
 import 'package:dartle/dartle_cache.dart';
 
 import 'config.dart';
-import 'config_source.dart';
 import 'jb_dartle.dart';
 import 'jb_files.dart';
 import 'jvm_executor.dart';
@@ -15,9 +14,8 @@ class JbRunner {
 
   JbRunner(this.files, this.config, this._jvmExecutor);
 
-  static Future<JbRunner> create(
-      JbFiles files, Sendable<JavaCommand, Object?> jvmExecutor) async {
-    final config = await _createConfig(files.configSource);
+  static Future<JbRunner> create(JbFiles files, JbConfiguration config,
+      Sendable<JavaCommand, Object?> jvmExecutor) async {
     logger.fine(() => 'Parsed jb configuration: $config');
     config.validate();
     return JbRunner(files, config, jvmExecutor);
@@ -34,17 +32,5 @@ class JbRunner {
     await jb.init;
 
     return await runBasic(jb.tasks, jb.defaultTasks, options, cache);
-  }
-}
-
-Future<JbConfiguration> _createConfig(ConfigSource configSource) async {
-  try {
-    return await configSource.load();
-  } on DartleException {
-    rethrow;
-  } catch (e) {
-    throw DartleException(
-        message: 'Unable to load jb config due to: $e.'
-            '\nRun with the --help option to see usage.');
   }
 }